connect websocket to Redis pub/sub
